fix(dashboard-management-cards): move components to general
