+++ src/components/Charts/BasicInfoCard/BasicInfoCard.tsx
@@ -1,7 +1,7 @@
 import { Text } from '@status-im/components'
 import { YStack } from 'tamagui'
 
-import DashboardCardWrapper from '../DashboardCardWrapper'
+import DashboardCardWrapper from '../../../pages/Dashboard/DashboardCardWrapper'
 
 type BasicInfoCardProps = {
   title: string
