+++ src/components/Charts/ConsensusUptime/ConsensusUptimeCard.tsx
@@ -1,25 +1,25 @@
 import { Text } from '@status-im/components'
 import { Stack, XStack, YStack } from 'tamagui'
 
-import Icon from '../../../components/General/Icon'
-import UptimeChart from '../UptimeChart/UptimeChart'
-import DashboardCardWrapper from '../DashboardCardWrapper'
+import UptimeChart from '../../../pages/Dashboard/UptimeChart/UptimeChart'
+import Icon from '../../General/Icon'
+import DashboardCardWrapper from '../../../pages/Dashboard/DashboardCardWrapper'
 
-const ExecutionUptime = () => {
+const ConsensusUptimeCard = () => {
   return (
     <DashboardCardWrapper style={{ height: '156px', minWidth: '200px' }}>
       <YStack space={'$3'}>
         <YStack>
           <Text size={15} weight={'semibold'}>
-            Execution Uptime
+            Consensus Uptime
           </Text>
           <XStack style={{ alignItems: 'end' }} space={'$2'}>
             <Text size={27} weight={'semibold'}>
-              99%
+              98%
             </Text>
             <Icon src="/icons/positive.svg" width={13} height={13}></Icon>
             <Text size={13} color="#E95460">
-              2%
+              1.56%
             </Text>
           </XStack>
         </YStack>
@@ -34,6 +34,10 @@ const ExecutionUptime = () => {
             }}
           >
             <UptimeChart
+              // monthlyActivity={monthlyActivity}
+              // startMonth={0}
+              // endMonth={1}
+              // withLabels={false}
               startMonth={0}
               endMonth={3}
               monthlyActivity={[10, 1, 3, 4, 5, 1, 7, 1, 6, 3, 1, 9]}
@@ -45,4 +49,4 @@ const ExecutionUptime = () => {
     </DashboardCardWrapper>
   )
 }
-export default ExecutionUptime
+export default ConsensusUptimeCard
