+++ src/components/Charts/ExecutionUptime/ExecutionUptime.tsx
@@ -1,25 +1,25 @@
 import { Text } from '@status-im/components'
 import { Stack, XStack, YStack } from 'tamagui'
 
-import UptimeChart from '../UptimeChart/UptimeChart'
-import Icon from '../../../components/General/Icon'
-import DashboardCardWrapper from '../DashboardCardWrapper'
+import Icon from '../../General/Icon'
+import UptimeChart from '../../../pages/Dashboard/UptimeChart/UptimeChart'
+import DashboardCardWrapper from '../../../pages/Dashboard/DashboardCardWrapper'
 
-const ConsensusUptimeCard = () => {
+const ExecutionUptime = () => {
   return (
     <DashboardCardWrapper style={{ height: '156px', minWidth: '200px' }}>
       <YStack space={'$3'}>
         <YStack>
           <Text size={15} weight={'semibold'}>
-            Consensus Uptime
+            Execution Uptime
           </Text>
           <XStack style={{ alignItems: 'end' }} space={'$2'}>
             <Text size={27} weight={'semibold'}>
-              98%
+              99%
             </Text>
             <Icon src="/icons/positive.svg" width={13} height={13}></Icon>
             <Text size={13} color="#E95460">
-              1.56%
+              2%
             </Text>
           </XStack>
         </YStack>
@@ -34,10 +34,6 @@ const ConsensusUptimeCard = () => {
             }}
           >
             <UptimeChart
-              // monthlyActivity={monthlyActivity}
-              // startMonth={0}
-              // endMonth={1}
-              // withLabels={false}
               startMonth={0}
               endMonth={3}
               monthlyActivity={[10, 1, 3, 4, 5, 1, 7, 1, 6, 3, 1, 9]}
@@ -49,4 +45,4 @@ const ConsensusUptimeCard = () => {
     </DashboardCardWrapper>
   )
 }
-export default ConsensusUptimeCard
+export default ExecutionUptime
