+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,11 +1,11 @@
 import { Stack, YStack, XStack } from 'tamagui'
 
-import BasicInfoCards from './BasicInfoCards/BasicInfoCards'
+import BasicInfoCards from './BasicInfoCards'
 import BalanceChartCard from './BalanceChartCard/BalanceChartCard'
 import CPUCard from './CPULoad/CPUCard'
-import ConsensusUptimeCard from './ConsensusUptime/ConsensusUptimeCard'
-import ExecutionUptime from './ExecutionUptime/ExecutionUptime'
-import DeviceUptime from './DeviceUptime/DeviceUptime'
+import ConsensusUptimeCard from '../../components/Charts/ConsensusUptime/ConsensusUptimeCard'
+import ExecutionUptime from '../../components/Charts/ExecutionUptime/ExecutionUptime'
+import DeviceUptime from '../../components/Charts/DeviceUptime/DeviceUptime'
 import NimbusLogo from '../../components/Logos/NimbusLogo'
 import StorageCard from './StorageCard/StorageCard'
 import NetworkCard from './NetworkCard/NetworkCard'
