+++ src/pages/ValidatorManagement/ValidatorManagement.tsx
@@ -2,10 +2,10 @@ import { Text } from '@status-im/components'
 import ManagementTabs from './ManagementTabs'
 import ManagementHeader from './ManagementHeader'
 import SidebarsWrapper from '../../components/PageWrappers/SidebarsWrapper'
-import DeviceUptime from '../Dashboard/DeviceUptime/DeviceUptime'
-import ConsensusUptimeCard from '../Dashboard/ConsensusUptime/ConsensusUptimeCard'
-import ExecutionUptime from '../Dashboard/ExecutionUptime/ExecutionUptime'
-import BasicInfoCard from '../Dashboard/BasicInfoCards/BasicInfoCard'
+import DeviceUptime from '../../components/Charts/DeviceUptime/DeviceUptime'
+import ConsensusUptimeCard from '../../components/Charts/ConsensusUptime/ConsensusUptimeCard'
+import ExecutionUptime from '../../components/Charts/ExecutionUptime/ExecutionUptime'
+import BasicInfoCard from '../../components/Charts/BasicInfoCard/BasicInfoCard'
 import styles from './validatorManagement.module.css'
 import { useWindowSize } from '../../hooks/useWindowSize'
